ENGR00156967 Deleted media file still can be seen in Gallery3D.
Some of the files thumbnails are not deleted. Fixed the bug by adding
the way to deal with deleting files in extsd or udisk.
But even when all of the media files are deleted, there still is one
media file can be seen in Gallery3D. It is a Gallery3D bug. Left it to
google to fix it. The issue in android bug report is 7827: Cannot delete
the last photo in the gallery.

Signed-off-by: YangYonggang <b31664@freescale.com>