ENGR00180234:fix connectivity priority bug
1.Data connectivity priority is Ethernet>wifi>3G
2.For ICS framework, ethernet is not considered in default.
With ethernet support,should add config item in value.xml.
For wifi is default as the high priority,so if change to
ethernet,should update setting database in ConnectivityService.
3.After unplug ethernet cable, ConnectivityService should auto
switch to wifi or 3G, but default framework don't trigger event
to ConnectivityService. this is because eth0 is only link down
instead of remove. ICS only trigger event if remove.

Signed-off-by: Jianzheng Zhou <B38613@freescale.com>