ENGR00215174 fix apk specify orientation different from system show incorrect.
reason:
when in dual display case, the UI does not rotate with sensor.
but apk can specify orientation different from system.
when do mirror function for secondary display, the rotation is
not taken into account.

resolve:
calculate the hardware layer position and size with rotation parameter.
when do mirror function.

Signed-off-by: Xiaowen Liu <b37945@freescale.com>