ENGR00315845 third device invitation impact WFD connection
When the third device try to connect the Sink or Source it will impact the
current WFD connect due to the frameworks issue. The third one use the p2p
or wfd connection will lead the GroupCreatedState exit to the
UserAuthizingJoinState. The that will cause the CONNECTION_CHANGED boardcast
which carries the DISCONNECT info. WifiP2pService will be told the WFD connection
state by the WifiDisplayController and the WfdSink. If it is a group of WFD
connection the state will not jump.

Signed-off-by: Wang Haoran <b50027@freescale.com>