PackageManager: add option to skip wifi direct feature
It has been noticed on TiWi-BLE module that Bluetooth wasn't performing
great when WiFi Direct feature is enabled on the platform. This patch is
a workaround to allow the user to disable the WiFi Direct without having
to re-build an image. It is based on the "persist.sys.wifidirect" property
which, if set to "disabled", discards the feature whether it is set or not
in /etc/permissions/. Platform reboot is required for the change to occur.

Change-Id: I60c3cc8015bb6c3831eee699c967e504b7f95c9b
Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>