fix starting window issue
Symptom:
1. Acitivity A and B belong to same task.
2. A is started and start B in its code
3. B is opened with blank screen

If starting icon has been display, both startingWindow and
startingView should not be null.
Current logic only uses judgement on startWindow. In above
case, it leads startingView set to null as A and B shares
the same starting window due to same task. Then system will
not get chance to remove startingView as it is null.

It needs add judgement on startingView.

Change-Id: I6cfa1a3ebce93ce435cc2d9ada239b1e808988de
Reviewed-on: http://androidsource.ap.freescale.net/project/79
Reviewed-by: Wang Haoran <elven.wang@nxp.com>